break `Combine` tests out into separate file, slash delays for shorter tests
